+++ src/16/outline/ChevronUp.js
@@ -1,0 +1,16 @@
+import * as React from 'react'
+const SvgChevronUp = (props) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={16}
+    height={16}
+    fill="currentColor"
+    {...props}
+  >
+    <path
+      fillRule="evenodd"
+      d="M3.235 10.795a.75.75 0 0 0 1.06-.03L8 6.842l3.705 3.923a.75.75 0 0 0 1.09-1.03l-4.25-4.5a.75.75 0 0 0-1.09 0l-4.25 4.5a.75.75 0 0 0 .03 1.06Z"
+    />
+  </svg>
+)
+export default SvgChevronUp