+++ src/16/solid/ChevronsLeft.js
@@ -1,0 +1,14 @@
+import * as React from 'react'
+const SvgChevronsLeft = (props) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={16}
+    height={16}
+    fill="currentColor"
+    {...props}
+  >
+    <path d="M7.76 5.3a.75.75 0 0 0-1.02-1.1l-3.5 3.25a.75.75 0 0 0 0 1.1l3.5 3.25a.749.749 0 1 0 1.02-1.1L4.852 8 7.76 5.3Z" />
+    <path d="M12.76 5.3a.75.75 0 0 0-1.02-1.1l-3.5 3.25a.75.75 0 0 0 0 1.1l3.5 3.25a.748.748 0 0 0 1.213-.29.75.75 0 0 0-.193-.81L9.852 8l2.908-2.7Z" />
+  </svg>
+)
+export default SvgChevronsLeft