+++ src/16/solid/ChevronsUp.js
@@ -1,0 +1,14 @@
+import * as React from 'react'
+const SvgChevronsUp = (props) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={16}
+    height={16}
+    fill="currentColor"
+    {...props}
+  >
+    <path d="M5.3 7.76a.75.75 0 0 1-1.1-1.02l3.25-3.5a.75.75 0 0 1 1.1 0l3.25 3.5a.749.749 0 1 1-1.1 1.02L8 4.852 5.3 7.76Z" />
+    <path d="M5.3 12.76a.75.75 0 0 1-1.1-1.02l3.25-3.5a.75.75 0 0 1 1.1 0l3.25 3.5a.748.748 0 0 1-.29 1.213.75.75 0 0 1-.81-.193L8 9.852 5.3 12.76Z" />
+  </svg>
+)
+export default SvgChevronsUp