+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.8] - 2023-11-18
+
+### Added
+
+- Added Chevron Right
+- Added Chevron Up
+- Added Chevrons Down
+- Added Chevrons Left
+- Added Chevrons Right
+- Added Chevrons Up
+
 ## [0.1.7] - 2023-11-12
 
 ### Fixed